-logtcp output to specified log file. clean up some warnings from w4 VS2005.

+++ dk_essentials.h
@@ -10,7 +10,7 @@ char *DK_strtok_r(char *s, const char *delim, char **last);
 char *Info_ValueForKey(const char *s, const char *key); // FS: From Quake 2
 void Gamespy_Create_Challenge_Key(char *s, const int len);
 const char *Gamespy_Get_Game_SecKey (char *gamename);
-int Gamespy_Get_MOTD_Port (char *gamename);
+unsigned short Gamespy_Get_MOTD_Port (char *gamename);
 char *DK_strlwr (char *s); // FS: Some compilers might not have this
 void Parse_ServerList (unsigned int fileSize, char *fileBuffer, char *gamenameFromHttp);
 void AddServers_From_List_Execute(char *fileBuffer, char *gamenameFromHttp); // FS: From Quake 2
